v1/requirements.txt
dspy-ai>=2.0.4
python-dotenv>=1.0.0
openai>=1.0.0
langchain-openai>=0.0.5
langchain-core>=0.1.15
langchain-community>=0.0.10
langgraph>=0.0.15
PyPDF2
faiss-cpu
pypdf
pdf2image
pdfplumber
PyMuPDF>=1.20.0  # fitz
Pillow>=10.0.0  # PIL
langchain>=0.1.0
python-dotenv>=1.0.0